Several updates to the disabled RHEL package list (#168)
Many of these packages were awaiting Bloom generation, which actually
happened a while ago. Some of these packages are also not currently
released for Galactic, so I'm not sure how they got on the list to begin
with.

For each package being enabled, I verified that it builds successfully,
but that doesn't mean that it won't unblock a downstream package that
does not.