Add alternative rmw to skip_rosdep_keys. (#178)
These dependencies are being resolved to rolling binary packages in the
underlay.

Another possible option, used by the other CI jobs, is to keep
everything in the repos file and use package selection arguments to
ignore the in-workspace dependencies. But that will result in our repos
file containing more than the minimum package set for the clang-tidy
target.