Update docs upload configuration.
In preparation for moving docs.ros.org to the new web host, we're
rolling out a new set of deploy keys as well as a dedicated user account
just for the docs.

This change needs to be merged in coordination with the Jenkins
deployment of the new credential and updated DNS for docs.ros.org as
these keys nor configuration are valid on the current docs host.